--- client/src/App.jsx
@@ -1,19 +1,0 @@
-
-
-import './App.css'
-import { Button } from 'antd';
-
-function App() {
-
-  return (
-    <>
-        <h1 className="text-3xl font-bold underline">
-      Hello world!
-    </h1>
-    <div></div>
-    <Button type="primary">Button</Button>
-    </>
-  )
-}
-
-export default App
