+++ client/tailwind.config.js
@@ -52,3 +52,4 @@ export default {
 };
 
 
+
